Add test cases for provisioning service client
